svcClient returns a structured object.
apiDefinitions can be grouped one level if desired, the return is an
object. So where before the interface was a higher order function that
took a method name ie. client('testMethod', params), that will now be
client.testMethod(params) which is a lot more intuitive for a typical
http client library.